815 refactor existing selected site metrics pane code into its own component
