+++ src/components/MetricsPane/MetricsPane.styles.js
@@ -2,16 +2,7 @@ import styled, { css } from 'styled-components'
 import { mediaQueryTabletLandscapeOnly, hoverState } from '../../styles/mediaQueries'
 import theme from '../../styles/theme'
 import { ButtonSecondary } from '../generic'
-import {
-  IconCaretUp,
-  IconCaretDown,
-  IconPersonCircle,
-  IconTextBoxMultiple,
-  IconCalendar,
-  IconText,
-} from '../../assets/dashboardOnlyIcons'
-import { IconGlobe, IconUser } from '../../assets/icons'
-import { Row } from '../generic/positioning'
+import { IconCaretUp, IconCaretDown } from '../../assets/dashboardOnlyIcons'
 
 export const StyledMetricsWrapper = styled.div`
   display: flex;
@@ -204,56 +195,6 @@ export const MobileExpandedMetricsPane = styled.div`
   width: 100%;
 `
 
-export const SelectedSiteMetricsCardContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  padding: 1rem;
-  border: 1px solid ${theme.color.grey0};
-  background-color: ${theme.color.white};
-`
-
-export const SelectedSiteSiteCardContainer = styled(SelectedSiteMetricsCardContainer)`
-  text-transform: uppercase;
-  align-items: center;
-`
-
-export const SelectedSiteContentContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 27.5rem;
-`
-
-export const biggerIcons = css`
-  width: ${theme.typography.mediumIconSize};
-  height: ${theme.typography.mediumIconSize};
-  position: relative;
-  margin-right: 0.8rem;
-`
-
-export const BiggerIconPersonCircle = styled(IconPersonCircle)`
-  ${biggerIcons}
-`
-
-export const BiggerIconTextBoxMultiple = styled(IconTextBoxMultiple)`
-  ${biggerIcons}
-`
-
-export const BiggerIconCalendar = styled(IconCalendar)`
-  ${biggerIcons}
-`
-
-export const BiggerIconUser = styled(IconUser)`
-  ${biggerIcons}
-`
-
-export const BiggerIconGlobe = styled(IconGlobe)`
-  ${biggerIcons}
-`
-
-export const BiggerIconText = styled(IconText)`
-  ${biggerIcons}
-`
-
 export const StyledHeader = styled.h2`
   font-size: ${theme.typography.defaultFontSize};
   font-weight: bold;
@@ -278,42 +219,6 @@ export const StyledMetricsSelector = styled.div`
   padding: 0.5rem;
 `
 
-export const StyledMapPinContainer = styled.div`
-  margin-right: 0.8rem;
-`
-
-export const StyledSvgContainer = styled.div`
-  width: 2.25rem;
-  height: 2.25rem;
-  margin-left: 0.2rem;
-  margin-right: 0.6rem;
-`
-
-export const StyledReefContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-`
-
-export const StyledReefRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-`
-export const StyledReefItem = styled.p`
-  width: 50%;
-  text-transform: capitalize;
-  margin: 0;
-`
-
-export const StyledReefItemBold = styled(StyledReefItem)`
-  font-weight: bold;
-  width: 50%;
-`
-export const StyledVisibleBackground = styled.div`
-  background-color: ${theme.color.grey1};
-`
-
 export const StyledChevronSpan = styled.span`
   padding-left: 0.8rem;
 `
@@ -334,12 +239,4 @@ export const DesktopFollowScreenButton = styled(ButtonSecondary)`
 
 export const StyledLabel = styled.label`
   cursor: pointer;
-`
-export const SelectedSiteActionBar = styled(Row)`
-  width: 100%;
-  gap: ${theme.spacing.small};
-
-  & > button {
-    width: 100%;
-  }
 `
